refactor(ui): Phase 1 — extract tool rendering into modular directory structure
Split the monolithic tool rendering system into organized modules:
- tool/registry.ts — ToolRegistry, ToolProps, ToolComponent (extracted from message-part.tsx)
- tool/classifier.ts — semantic classifier (moved from semantic-tool-classifier.ts)
- tool/body-primitives.tsx — SummaryGrid, WarningPanel, DiagnosticsPanel, RawOutput, SearchFiles + helpers (extracted from anchored-tool-card.tsx)
- tool/renders/file-ops.tsx — 8 file operation tools (view_file, scan_files, parse_code, revise_file, save_file, edit, write, multiedit)
- tool/renders/standard.tsx — 48 BasicTool+trigger pattern tools
- tool/renders/task.tsx — task tool with subagent summary
- tool/renders/dag.tsx — dagwrite/dagpatch/dagread with DAG graph
- tool/renders/special.tsx — diagram tool with DiagramRenderer
- tool/renders/batch.tsx — qzcli/inspire/research/worktree batch registrations
- tool/renders/index.ts — barrel export

Backward compatibility preserved: tool-renders.tsx is now a thin re-export of all render files. session-turn.tsx import untouched. All original rendering logic preserved. Typecheck: 8/8 pass.

Part of tool card rendering architecture overhaul.
