+++ packages/ui/src/components/tool/renders/dag.tsx
@@ -1,0 +1,119 @@
+import { Show } from "solid-js"
+import { BasicTool } from "../../basic-tool"
+import { DagGraph } from "../../dag-graph"
+import { ToolRegistry } from "../../message-part"
+import { getSemanticIcon } from "../../semantic-icon"
+
+ToolRegistry.register({
+  name: "dagwrite",
+  render(props) {
+    const nodes = () =>
+      (props.metadata?.nodes ?? props.input?.nodes ?? []) as {
+        id: string
+        content: string
+        status: string
+        deps: string[]
+        assign?: string
+      }[]
+    const ready = () => (props.metadata?.ready ?? []) as string[]
+    const completed = () => nodes().filter((n) => n.status === "completed").length
+    const total = () => nodes().length
+    const firstReady = () => {
+      const readyIds = ready()
+      if (readyIds.length > 0) {
+        const node = nodes().find((n) => n.id === readyIds[0])
+        if (node) {
+          const text = node.content ?? ""
+          return text.length > 30 ? text.slice(0, 27) + "…" : text
+        }
+      }
+      const pending = nodes().find((n) => n.status === "pending" || n.status === "running")
+      if (pending) {
+        const text = pending.content ?? ""
+        return text.length > 30 ? text.slice(0, 27) + "…" : text
+      }
+      return ""
+    }
+    const ratio = () => (total() > 0 ? `${completed()}/${total()}` : "")
+    return (
+      <BasicTool
+        {...props}
+        icon={getSemanticIcon("orchestration.dag")}
+        trigger={() => ({
+          title: "DAG",
+          subtitle: firstReady() || "",
+          args: ratio() ? [ratio()] : [],
+        })}
+      >
+        <Show when={(nodes()?.length ?? 0) > 0}>
+          <DagGraph nodes={nodes()} ready={ready()} />
+        </Show>
+      </BasicTool>
+    )
+  },
+})
+
+ToolRegistry.register({
+  name: "dagpatch",
+  render(props) {
+    const nodes = () =>
+      (props.metadata?.nodes ?? []) as {
+        id: string
+        content: string
+        status: string
+        deps: string[]
+        assign?: string
+      }[]
+    const ready = () => (props.metadata?.ready ?? []) as string[]
+    const completed = () => nodes().filter((n) => n.status === "completed").length
+    const total = () => nodes().length
+    const ratio = () => (total() > 0 ? `${completed()}/${total()}` : "")
+    return (
+      <BasicTool
+        {...props}
+        icon={getSemanticIcon("orchestration.dag")}
+        trigger={() => ({
+          title: "DAG",
+          subtitle: "updated",
+          args: ratio() ? [ratio()] : [],
+        })}
+      >
+        <Show when={(nodes()?.length ?? 0) > 0}>
+          <DagGraph nodes={nodes()} ready={ready()} />
+        </Show>
+      </BasicTool>
+    )
+  },
+})
+
+ToolRegistry.register({
+  name: "dagread",
+  render(props) {
+    const nodes = () =>
+      (props.metadata?.nodes ?? []) as {
+        id: string
+        content: string
+        status: string
+        deps: string[]
+        assign?: string
+      }[]
+    const ready = () => (props.metadata?.ready ?? []) as string[]
+    const completed = () => nodes().filter((n) => n.status === "completed").length
+    const total = () => nodes().length
+    const ratio = () => (total() > 0 ? `${completed()}/${total()}` : "")
+    return (
+      <BasicTool
+        {...props}
+        icon={getSemanticIcon("orchestration.dag")}
+        trigger={() => ({
+          title: "Read DAG",
+          args: ratio() ? [ratio()] : [],
+        })}
+      >
+        <Show when={(nodes()?.length ?? 0) > 0}>
+          <DagGraph nodes={nodes()} ready={ready()} />
+        </Show>
+      </BasicTool>
+    )
+  },
+})
